requirements.txt
Pillow=9.4.0
mouseinfo=0.1.3
pynput=1.7.6
